Buzzer

+++ include/pwm.h
@@ -1,4 +1,6 @@
 #include <stdint.h>
 
+//method to initilaise pwm
 void pwm_pwm_init();
-void brightness_adjust(uint16_t input);
+//method to adjust the brightness
+void brightness_adjust(uint8_t input);
